sbi: Move attestation to its own file
Continue reducing clutter in sbi.rs
